podcast-script/requirements.txt
anthropic
feedparser
openai>=1.0
python-dateutil
python-dotenv
